video: break hard dependency on dnn
