+++ modules/dnn_algorithms/include/opencv2/dnn_algorithms.hpp
@@ -1,0 +1,12 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+
+#ifndef OPENCV_DNN_ALGORITHMS_HPP
+#define OPENCV_DNN_ALGORITHMS_HPP
+
+#include "opencv2/dnn_algorithms/face.hpp"
+#include "opencv2/dnn_algorithms/tracking.hpp"
+#include "opencv2/dnn_algorithms/super_scale.hpp"
+
+#endif
