+++ modules/dnn_algorithms/misc/js/gen_dict.json
@@ -1,0 +1,6 @@
+{
+    "whitelist":
+    {
+        "FaceDetectorYN": ["setInputSize", "getInputSize", "setScoreThreshold", "getScoreThreshold", "setNMSThreshold", "getNMSThreshold", "setTopK", "getTopK", "detect", "create"]
+    }
+}
