+++ modules/dnn_algorithms/perf/perf_main.cpp
@@ -1,0 +1,7 @@
+#include "perf_precomp.hpp"
+
+#if defined(HAVE_HPX)
+    #include <hpx/hpx_main.hpp>
+#endif
+
+CV_PERF_TEST_MAIN(dnn_algorithms)
