+++ modules/dnn_algorithms/perf/perf_precomp.hpp
@@ -1,0 +1,7 @@
+#ifndef __OPENCV_PERF_PRECOMP_HPP__
+#define __OPENCV_PERF_PRECOMP_HPP__
+
+#include "opencv2/ts.hpp"
+#include "opencv2/dnn_algorithms.hpp"
+
+#endif
