+++ modules/dnn_algorithms/src/precomp.hpp
@@ -1,0 +1,18 @@
+// This file is part of OpenCV project.
+// It is subject to the license terms in the LICENSE file found in the top-level directory
+// of this distribution and at http://opencv.org/license.html.
+
+#ifndef DNN_ALGORITHMS_PRECOMP_HPP
+#define DNN_ALGORITHMS_PRECOMP_HPP
+
+#include "opencv2/core/utils/logger.hpp"
+
+#include "opencv2/dnn_algorithms.hpp"
+#include "opencv2/core.hpp"
+#include "opencv2/imgproc.hpp"
+#include "opencv2/dnn.hpp"
+#include "opencv2/video/tracking.hpp"
+
+#include <algorithm>
+
+#endif
