+++ modules/dnn_algorithms/test/test_precomp.hpp
@@ -1,15 +1,12 @@
 // This file is part of OpenCV project.
 // It is subject to the license terms in the LICENSE file found in the top-level directory
 // of this distribution and at http://opencv.org/license.html.
-#include "test_precomp.hpp"
+#ifndef __OPENCV_TEST_PRECOMP_HPP__
+#define __OPENCV_TEST_PRECOMP_HPP__
 
-#if defined(HAVE_HPX)
-    #include <hpx/hpx_main.hpp>
+#include "opencv2/ts.hpp"
+#include "opencv2/dnn_algorithms.hpp"
+
+#include <random>
+
 #endif
-
-static
-void initTests()
-{
-}
-
-CV_TEST_MAIN("cv", initTests())
