+++ modules/objdetect/include/opencv2/objdetect.hpp
@@ -866,7 +866,6 @@ public:
 }
 
 #include "opencv2/objdetect/detection_based_tracker.hpp"
-#include "opencv2/objdetect/face.hpp"
 #include "opencv2/objdetect/charuco_detector.hpp"
 #include "opencv2/objdetect/barcode.hpp"
