+++ samples/dnn/face_detect.cpp
@@ -1,7 +1,7 @@
 #include <opencv2/dnn.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/objdetect.hpp>
+#include <opencv2/dnn_algorithms.hpp>
 
 #include <iostream>
