+++ samples/dnn/nanotrack_tracker.cpp
@@ -10,7 +10,7 @@
 #include <opencv2/dnn.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/video.hpp>
+#include <opencv2/dnn_algorithms.hpp>
 
 using namespace cv;
 using namespace cv::dnn;
