+++ samples/dnn/vit_tracker.cpp
@@ -7,7 +7,7 @@
 #include <opencv2/dnn.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/video.hpp>
+#include <opencv2/dnn_algorithms.hpp>
 
 using namespace cv;
 using namespace cv::dnn;
